First version of permission modal for devices
Create a first version of a permission modal for managing permissions of devices.
So far, it allows to enable and modify public link and includes a mockup for individual user permissions.

For publick link, it has a toggle near the title to activate or deactivate it.
When activated, an input for modifying the link is shown.
A button is also shown which allows you to directly share the full link.

Regarding the individual user permissions, they are represented as chips.
The chips contain an avatar image to the left and edit icon to the right.
When the edit icon is shown, it shall allow to modify permissions, although it is not yet implemented.
For future, clicking the chip will open the user info as a modal.

Ion Notes have been used for section titles.