Redesign device sharing logic
The device sharing (on the _MORE_ action sheet), is completely
redesigned to a now working method.
Earlier on, it did not properly work on all devices, and it did not
correctly share the generated JSON file.

The new method is based on the device public link.
Now, the device public link is what actually gets shared (what you could
previously only do through the premissions modal), and the share
Capacitor API is called on this link.
When the Share API is not available on some OS's and browsers, a
fallback is ran, which copies the link to the clipboard after getting
user confirmation.

The next step is placing a toast to confirm that the link was copied to
the clipboard.