Show clipboard copy alert only when needed
Code has been modified so clipboard copy only shows when Share API is
not available, instead when it is available too.
This change is due to a experience where the API was available, when
when discarting the share dialog, the clipboard alert appeared.
This simply shows the alert only if the error is the one raised when the
API is not available.