Fix (finally) undefined device action sheet
The fix is actually quite simple for a so long annoying bug.
The problem was that the ellipsis icon was also calling the
`moreActionSeet` action, but without passing the device.
This caused the action sheet to appear twice and the device undefined.
This is fixed by simply removing the `@click` bind for the icon.